requirements.txt
# cat requirements.txt

django=2.0.5
django-activity-stream
django-notifications-hq